Add from_chars for integers :1234:
After a lot of experimenting with sse2/ssse3 simd extensions, and
comparing with non-vectorised code, it can be concluded that any simd
algorithm at best (i.e. an implementation that disregards all error
checking) will still always be slower for numbers with few digits and
only noticeably faster with 6 digits or so. Adding overhead to this,
whether it comes from fixing the issue of few digits or adding error
checking to make the function usable in practice, makes the algorithm
slow enough to not be worth the effort. A simple implementation is
therefore the best.